source-chooser: fix "TypeError: filter is not a function" exception
When clicking the source chooser button a "TypeError: filter is not a function"
error message appears in the console.

This error appears to be caused by the conversion of mejs.Utils away from jQuery
which used to allow mejs.Utils.siblings() to take a class selector string: the
current implementation requires the second parameter to be a filter function.